+++ include/tokenizer.h
@@ -3,6 +3,7 @@
 
 #include "minishell.h"
 
+t_tokenizer	make_tokenizer(char *line);
 t_tokenizer	get_next_token(t_tokenizer *tokenizer);
 t_tokenizer	scan_char_token(t_tokenizer *tokenizer);
 t_tokenizer	make_token(t_tokenizer *tokenizer, t_symbol type);
@@ -12,5 +13,7 @@ void		reset_tokenizer(t_tokenizer *tokenizer);
 void		skip_whitespace(t_tokenizer *token);
 t_bool		match(t_tokenizer *tokenizer, char matchword);
 t_bool		string_close(t_tokenizer *tokenizer, char c);
+
+ t_node	    *parse(t_tokenizer *token);
 
 #endif
